server support for loading friend network info (for hovercard) from site
