+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/core/NetworkInfoLoader.scala
@@ -1,0 +1,48 @@
+package com.keepit.controllers.core
+
+import com.google.inject.{Inject, Singleton}
+import com.keepit.common.db._
+import com.keepit.common.db.slick._
+import com.keepit.common.social._
+import com.keepit.social.{SocialNetworkType, SocialId}
+import com.keepit.model._
+
+import play.api.libs.functional.syntax._
+import play.api.libs.json._
+
+case class NetworkInfo(profileUrl: Option[String], connected: Boolean)
+
+object NetworkInfo {
+  implicit val writesNetworkInfo = Json.writes[NetworkInfo]
+  implicit val writesNetworkTypeToNetworkInfo = new Writes[Map[SocialNetworkType, NetworkInfo]] {
+    def writes(o: Map[SocialNetworkType, NetworkInfo]): JsValue =
+      JsObject(o.map { case (network, info) => network.name -> Json.toJson(info) }.toSeq)
+  }
+}
+
+@Singleton
+class NetworkInfoLoader @Inject() (
+  db: Database,
+  userRepo: UserRepo,
+  socialConnectionRepo: SocialConnectionRepo,
+  socialUserInfoRepo: SocialUserInfoRepo) {
+
+  def load(userId: Id[User], friendId: ExternalId[User]): Map[SocialNetworkType, NetworkInfo] = {
+    db.readOnly { implicit s =>
+      val mySocialUsers = socialUserInfoRepo.getByUser(userId)
+      for {
+        friend <- userRepo.getOpt(friendId).toSeq
+        su <- socialUserInfoRepo.getByUser(friend.id.get)
+      } yield {
+        su.networkType -> NetworkInfo(
+          profileUrl = su.getProfileUrl,
+          connected = mySocialUsers.exists { mySu =>
+            mySu.networkType == su.networkType &&
+                socialConnectionRepo.getConnectionOpt(su.id.get, mySu.id.get).isDefined
+          }
+        )
+      }
+    }.toMap
+  }
+
+}
